- support of function declaration and definition;
- support for native functions;
- added 'Call frame' paradigm;

TODO:
    1. Refactor the structure of .fvb object files, because its current represenation leads to the following issues:
        - no globals at runtime;
        - no strings at runtime;
    2. Implement arrays;
    3. if-(else if)-else statement;
    4. Ternary operator support (challenge #3, chapter. 17);
    5. 'increment' and 'decrement' statements;
    6. Enhanced assignment, i.e. '+=', '*=', '-=' and '/=';
    7. 'switch-case', 'continue' and 'break';
    8. try-catch block;
    9. Concatenation of different types (challenge #3, chapter 19);
    10. String interpolation support (challenge #1, chapter. 16);
    11. Fully-featured Hash Table, which shall support not only strings, but all other predefined
        and user-defined types;
    12. The 'const' modifier;
    13. Flexible array members (challenge #1, chapter 19);
    14. Refactor and optimize the run() function;
    15. Advanced support for native functions;

+++ common/value.h
@@ -17,7 +17,7 @@ typedef struct {
 	ValueType type;
 	union {
 		bool boolean;
-		int32_t number;
+		float number;
 		Obj* obj;
 	} as;
 } Value;
